+++ src/diffusers/utils/dynamic_slice.py
@@ -1,0 +1,10 @@
+from torch import Tensor
+from typing import List
+
+def dynamic_slice(
+    x: Tensor,
+    starts: List[int],
+    sizes: List[int],
+) -> Tensor:
+    slicing = [slice(start, start + size + 1) for start, size in zip(starts, sizes)]
+    return x[slicing]
